nyc and deps, code coverage, npm scripts. linting, moving types for test coverage exclusion. test config tweaks.
