+++ .gitignore
@@ -2,3 +2,5 @@ dist
 node_modules
 .idea
 .DS_Store
+.nyc_output
+coverage
