+++ package.json
@@ -14,11 +14,12 @@
     "prod:start": "NODE_ENV=production NODE_OPTIONS='--import tsx' npx knex migrate:latest && node ./src/index.ts",
     "test:migrate": "NODE_ENV=test NODE_OPTIONS='--import tsx' npx knex migrate:latest",
     "test:seed": "NODE_ENV=test NODE_OPTIONS='--import tsx' npx knex seed:run",
-    "test:unit": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/unit/*.test.ts",
-    "test:integration-db": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json --recursive ./test/integration/db/*.test.ts",
-    "test:integration-api": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json --recursive ./test/integration/api/*.test.ts",
-    "test:integration": "npm run test:integration-db && npm run test:integration-api",
-    "test": "npm run test:unit && npm run test:integration"
+    "test:unit": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/unit/",
+    "test:integration-db": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/integration/db/",
+    "test:integration-api": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/integration/api/",
+    "test:integration": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/integration/",
+    "test": "NODE_ENV=test NODE_OPTIONS='--import tsx' mocha --config ./test/mocharc.json ./test/",
+    "test:coverage": "NODE_ENV=test NODE_OPTIONS='--import tsx' nyc mocha --config ./test/mocharc.json ./test/"
   },
   "engines": {
     "node": ">=20.11"
@@ -51,6 +52,7 @@
     "tsx": "^4.7.1"
   },
   "devDependencies": {
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@types/chai": "^4.3.14",
     "@types/mocha": "^10.0.6",
     "@types/supertest": "^6.0.2",
@@ -64,6 +66,7 @@
     "eslint-plugin-promise": "^6.1.1",
     "mocha": "^10.3.0",
     "nodemon": "^3.1.0",
+    "nyc": "^15.1.0",
     "supertest": "^6.3.4"
   }
 }
