+++ test/helper/thingRouteHelper.ts
@@ -1,8 +1,8 @@
 import request from 'supertest'
+import { Express } from 'express'
 
 import env from '../../src/env'
-import { Express } from 'express'
-import { ThingPayload, ThingType } from '../../src/types'
+import { Thing, ThingGroup, ThingPayload, ThingType } from '../../src/types/types'
 
 export const getThingTypesRoute = async (app: Express): Promise<any> => {
   return request(app)
@@ -77,7 +77,7 @@ export const getThingGroupDeviceByNameAndDeviceIdRoute = async (
     })
 }
 
-export const getThingGroupDevicesByNameRoute = async (app: Express, name?: string | unknown): Promise<any> => {
+export const getThingGroupDevicesByNameRoute = async (app: Express, name: unknown): Promise<any> => {
   return request(app)
     .get(`/${env.API_MAJOR_VERSION}/thingGroup/${name}/thing`)
     .set('content-type', 'application/json')
@@ -109,11 +109,11 @@ export const postThingGroupDeviceRoute = async (
     })
 }
 
-export const postThingGroupRoute = async (app: Express, thingGroup?: Record<string, unknown>): Promise<any> => {
+export const postThingGroupRoute = async (app: Express, thingGroup?: unknown): Promise<any> => {
   return request(app)
     .post(`/${env.API_MAJOR_VERSION}/thingGroup`)
     .set('content-type', 'application/json')
-    .send(thingGroup)
+    .send(thingGroup as ThingGroup)
     .then((response) => {
       return response
     })
@@ -124,11 +124,11 @@ export const postThingGroupRoute = async (app: Express, thingGroup?: Record<stri
     })
 }
 
-export const postThingTypeRoute = async (app: Express, thingType: ThingType): Promise<any> => {
+export const postThingTypeRoute = async (app: Express, thingType?: unknown): Promise<any> => {
   return request(app)
     .post(`/${env.API_MAJOR_VERSION}/thingType`)
     .set('content-type', 'application/json')
-    .send(thingType)
+    .send(thingType as ThingType)
     .then((response) => {
       return response
     })
@@ -167,11 +167,11 @@ export const getThingsRoute = async (app: Express): Promise<any> => {
     })
 }
 
-export const postThingRoute = async (app: Express, thing?: Record<string, unknown>): Promise<any> => {
+export const postThingRoute = async (app: Express, thing?: unknown): Promise<any> => {
   return request(app)
     .post(`/${env.API_MAJOR_VERSION}/thing`)
     .set('content-type', 'application/json')
-    .send(thing)
+    .send(thing as Thing)
     .then((response) => {
       return response
     })
@@ -182,7 +182,7 @@ export const postThingRoute = async (app: Express, thing?: Record<string, unknow
     })
 }
 
-export const postThingPayloadRoute = async (app: Express, thingPayload: any): Promise<any> => {
+export const postThingPayloadRoute = async (app: Express, thingPayload: unknown): Promise<any> => {
   return request(app)
     .post(`/${env.API_MAJOR_VERSION}/thingPayload`)
     .set('content-type', 'application/json')
