+++ tsconfig.json
@@ -14,8 +14,6 @@
     "isolatedModules": true,
     "resolveJsonModule": true
   },
-  "include": [
-    "./**/*.ts"
-  ],
+  "include": ["./**/*.ts"],
   "exclude": ["node_modules"]
 }
